Move BufferSource into Language and rename it to LanguageSource
